ajustes finais para manter atualizado com o FreeRTOS
Make changes necessary to allow the latest StateViewer Eclipse plug-in
obtain the run time stats information.

Default configUSE_TICKLESS_IDLE to 0 when it is not defined.
Move location of traceTASK_CREATE() macro call.

Remove obsolete handling of trmCOMMAND_PROCESS_TIMER_OVERFLOW as the
command was never used in release versions.

Move the call that sets xTimeNow inside the loop that drains the timer
queue to ensure higher priority tasks that pre-empt the timer daemon
cannot post messages that appear to be in the future to the daemon task.