Auxiliary function defined as the Task naked and always_inline The
Once the function is always_inline it need not be noreturn not naked.

function used in Infinite Loop task is just an object to facilitate the
use of the macro that facilitates the access to Task structure for use
in the FreeRTOS, so this function must be inserted into the code so
which are optimum and avoiding headers function.